se comenzo la clase inscripcionData
se realiza el metodo guardarinsscripcion

+++ src/universidadgrupo49/Entidades/Inscripcion.java
@@ -48,7 +48,7 @@ public class Inscripcion {
         this.materia = materia;
     }
 
-    public double getNota() {
+    public int getNota() {
         return nota;
     }
